+++ game/mod_sdk/materials/models/buggy/ammo_box.vmt
@@ -1,0 +1,10 @@
+"VertexLitGeneric"
+{
+	"$basetexture" "Models/Buggy/ammo_box"
+	"$surfaceprop" "metal"
+	"$envmap" "env_cubemap"
+ 	"$basealphaenvmapmask" 1
+	"$envmaptint" "[ .7 .7 .7]"
+}
+
+
