+++ game/mod_sdk/materials/models/buggy/buggy001.vmt
@@ -1,0 +1,9 @@
+"VertexLitGeneric"
+{
+	"$basetexture" "Models/Buggy/buggy001"
+	"$surfaceprop" "metal"
+	"$envmap" "env_cubemap"
+ 	"$basealphaenvmapmask" 1
+	"$envmaptint" "[ .5 .5 .5]"
+	"360?$color2" "[ 0.8 0.8 0.8 ]"
+}
