+++ src/game/client/client_base.vpc
@@ -222,7 +222,6 @@ $Project
 		$File	"c_test_proxytoggle.cpp"
 		$File	"c_user_message_register.cpp"
 		$File	"c_vehicle_choreo_generic.cpp"
-		$File	"c_vehicle_jeep.cpp"
 		$File	"c_vguiscreen.cpp"
 		$File	"c_waterbullet.cpp"
 		$File	"C_WaterLODControl.cpp"
@@ -635,7 +634,6 @@ $Project
 		$File	"c_tesla.h"
 		$File	"c_tracer.h"
 		$File	"c_triggers.h"
-		$File	"c_vehicle_jeep.h"
 		$File	"c_user_message_register.h"
 		$File	"c_vguiscreen.h"
 		$File	"c_weapon__stubs.h"
